Inline Cases enum and inline other external test code into the test classes (#6107)
