+++ detekt-rules-complexity/src/test/kotlin/io/gitlab/arturbosch/detekt/rules/complexity/LargeClassSpec.kt
@@ -1,6 +1,5 @@
 package io.gitlab.arturbosch.detekt.rules.complexity
 
-import io.github.detekt.test.utils.resourceAsPath
 import io.gitlab.arturbosch.detekt.api.SourceLocation
 import io.gitlab.arturbosch.detekt.test.TestConfig
 import io.gitlab.arturbosch.detekt.test.assertThat
@@ -13,10 +12,34 @@ private fun subject(threshold: Int) = LargeClass(TestConfig("threshold" to thres
 class LargeClassSpec {
 
     @Test
-    fun `should detect only the nested large class which exceeds threshold 70`() {
-        val findings = subject(threshold = 70).lint(resourceAsPath("NestedClasses.kt"))
+    fun `should detect only the nested large class which exceeds the threshold`() {
+        val code = """
+            class NestedClasses {
+            
+                private val i = 0
+            
+                class InnerClass {
+            
+                    class NestedInnerClass {
+            
+                        fun nestedMethod() {
+                            fun nestedLocalMethod() {
+                                println()
+                            }
+                            nestedLocalMethod()
+                        }
+                    }
+                }
+            }
+            
+            /**
+             * Top level members must be skipped for LargeClass rule
+             */
+            val aTopLevelPropertyOfNestedClasses = 0
+        """.trimIndent()
+        val findings = subject(threshold = 4).lint(code)
         assertThat(findings).hasSize(1)
-        assertThat(findings).hasStartSourceLocations(SourceLocation(12, 15))
+        assertThat(findings).hasStartSourceLocations(SourceLocation(7, 15))
     }
 
     @Test
