+++ detekt-rules-complexity/src/test/kotlin/io/gitlab/arturbosch/detekt/rules/complexity/NestedBlockDepthSpec.kt
@@ -1,6 +1,5 @@
 package io.gitlab.arturbosch.detekt.rules.complexity
 
-import io.github.detekt.test.utils.resourceAsPath
 import io.gitlab.arturbosch.detekt.api.ThresholdedCodeSmell
 import io.gitlab.arturbosch.detekt.test.TestConfig
 import io.gitlab.arturbosch.detekt.test.assertThat
@@ -16,8 +15,48 @@ class NestedBlockDepthSpec {
     val subject = NestedBlockDepth(defaultConfig)
 
     @Test
-    fun `should detect only the nested large class`() {
-        subject.lint(resourceAsPath("NestedClasses.kt"))
+    fun `should ignore class nesting levels`() {
+        val code = """
+            class NestedClasses {
+            
+                class InnerClass {
+            
+                    class NestedInnerClass {
+            
+                        fun nestedLongMethod() {
+                            if (true) {
+                                if (true) {
+                                    if (true) {
+                                        5.run {
+                                            this.let {
+                                                listOf(1, 2, 3).map { it * 2 }
+                                                    .groupBy(Int::toString, Int::toString)
+                                            }
+                                        }
+                                    }
+                                }
+                            }
+            
+                            try {
+                                for (i in 1..5) {
+                                    when (i) {
+                                        1 -> print(1)
+                                    }
+                                }
+                            } finally {
+            
+                            }
+            
+                            fun nestedLocalMethod() {
+                                println()
+                            }
+                            nestedLocalMethod()
+                        }
+                    }
+                }
+            }
+        """.trimIndent()
+        subject.lint(code)
         assertThat(subject.findings).hasSize(1)
         assertThat((subject.findings[0] as ThresholdedCodeSmell).value).isEqualTo(5)
     }
