--- detekt-rules-style/src/test/resources/NoTabsNegative.kt
@@ -1,11 +1,0 @@
-package cases
-
-class NoTabsNegative {
-
-    fun methodOk() {
-        println("A message")
-    }
-
-    val str = "A \t tab	"
-    val multiStr = """A \t tab	"""
-}
