--- detekt-rules-style/src/test/resources/NoTabsPositive.kt
@@ -1,11 +1,0 @@
-package cases
-
-class NoTabsPositive {
-	fun methodOk() { // reports 3
-		println("A message")
-
-	}
-
-  val str = "${		methodOk()}" // reports 1
-  val multiStr = """${	methodOk()}""" // reports 1
-}
