--- detekt-rules/src/test/resources/SuppressStringLiteralDuplication.kt
@@ -1,9 +1,0 @@
-@file:Suppress("unused", "UNUSED_PARAMETER", "UnusedEquals")
-
-@Suppress("StringLiteralDuplication")
-class Duplication {
-    var s1 = "lorem"
-    fun f(s: String = "lorem") {
-        s1 == "lorem"
-    }
-}
